requirements.txt
streamlit
plotly
pytz
streamlit-option-menu
streamlit
streamlit-option-menu
pandas
numpy
matplotlib
seaborn
scikit-learn
openpyxl
tabulate
Pillow
